RE-1317 Add function to abstract standard jobs
The pre-merge and post-merge standard jobs mostly duplicate each others
code. Soon there will be a final integration standard job, which will
mostly be another copy of the same code. This commit seeks to pull out
the standard elements into a separate function to simplify the
maintenance of the code and remove the need to copy/paste when creating
a new standard job.